# server/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "server project",
        "title": "Swagger server API",
        "contact": {
            "name": "antgo",
            "url": "https://github.com/small-ek/antgo",
            "email": "56494565@qq.com"
        },
        "version": "1.0"
    },
    "host": "127.0.0.1:9001",
    "basePath": "/api",
    "paths": {
...
        "/sys-menu": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "后台菜单"
                ],
                "summary": "获取后台菜单分页数据",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "当前页",
                        "name": "current_page",
                        "in": "query"
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "boolean"
                        },
                        "collectionFormat": "csv",
                        "description": "排序方式 true 降序 false 升序",
                        "name": "desc",
                        "in": "query"
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "collectionFormat": "csv",
                        "description": "过滤条件",
                        "name": "filter",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "图标",
                        "name": "icon",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "标识",
                        "name": "id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "忽略字段",
                        "name": "omit",
                        "in": "query"
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "collectionFormat": "csv",
                        "description": "排序字段",
                        "name": "order",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "每页显示数量",
                        "name": "page_size",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "父级标识",
                        "name": "parent_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "菜单跳转地址",
                        "name": "path",
                        "in": "query"
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "collectionFormat": "csv",
                        "description": "选择字段",
                        "name": "select",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "排序",
                        "name": "sort",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "是否隐藏0=禁用;1=启用",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "菜单名称",
                        "name": "title",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Write"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "allOf": [
                                                {
                                                    "$ref": "#/definitions/response.Page"
                                                },
                                                {
                                                    "type": "object",
                                                    "properties": {
                                                        "items": {
                                                            "type": "array",
                                                            "items": {
                                                                "$ref": "#/definitions/models.SysMenu"
                                                            }
                                                        }
                                                    }
                                                }
                                            ]
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/response.Write"
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "后台菜单"
                ],
                "summary": "创建后台菜单数据",
                "parameters": [
                    {
                        "description": "创建参数",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.SysMenuRequestForm"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Write"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/response.Write"
                        }
                    }
                }
            }
        },
        "/sys-menu/:id": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "后台菜单"
                ],
                "summary": "获取后台菜单详情数据",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Write"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/models.SysMenu"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/response.Write"
                        }
                    }
                }
            },
            "put": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "后台菜单"
                ],
                "summary": "修改后台菜单数据",
                "parameters": [
                    {
                        "description": "更新参数",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.SysMenuRequestForm"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Write"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/response.Write"
                        }
                    }
                }
            },
            "delete": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "后台菜单"
                ],
                "summary": "删除后台菜单数据",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Write"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/response.Write"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "models.SysMenu": {
            "type": "object",
            "properties": {
                "icon": {
                    "description": "图标",
                    "type": "string"
                },
                "id": {
                    "description": "标识",
                    "type": "integer"
                },
                "parent_id": {
                    "description": "父级标识",
                    "type": "integer"
                },
                "path": {
                    "description": "菜单跳转地址",
                    "type": "string"
                },
                "sort": {
                    "description": "排序",
                    "type": "integer"
                },
                "status": {
                    "description": "是否隐藏0=禁用;1=启用",
                    "type": "integer"
                },
                "title": {
                    "description": "菜单名称",
                    "type": "string"
                }
            }
        },
...
        "request.SysMenuRequestForm": {
            "type": "object",
            "required": [
                "title"
            ],
            "properties": {
                "icon": {
                    "description": "图标",
                    "type": "string"
                },
                "id": {
                    "description": "标识",
                    "type": "integer"
                },
                "parent_id": {
                    "description": "父级标识",
                    "type": "integer"
                },
                "path": {
                    "description": "菜单跳转地址",
                    "type": "string"
                },
                "sort": {
                    "description": "排序",
                    "type": "integer"
                },
                "status": {
                    "description": "是否隐藏0=禁用;1=启用",
                    "type": "integer"
                },
                "title": {
                    "description": "菜单名称",
                    "type": "string"
                }
            }
        },
        "response.Page": {
            "type": "object",
            "properties": {
                "items": {
                    "description": "Data set of the current page 数据集合"
                },
                "total": {
                    "description": "Total pages 总数量",
                    "type": "integer"
                }
            }
        },
        "response.Write": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "Status Code 业务逻辑的状态码",
                    "type": "integer"
                },
                "data": {
                    "description": "Data 实际数据"
                },
                "error": {
                    "description": "Error message 错误相关的信息"
                },
                "message": {
                    "description": "Msg Prompt message 简短描述",
                    "type": "string"
                }
            }
        },
...
    }
}